requirements/dev.txt
-r base.txt

pydevd-pycharm~=203.6682.179
django-debug-toolbar
